+++ include/globalvariables.h
@@ -1,0 +1,13 @@
+#pragma once
+
+namespace EquationData
+{
+const double perm = 1e-9;
+const double eta = 1;
+const double kappa = 1e-6;
+const double beat = 10;
+const double density = 1;
+const double T0 = 278.15;
+const double p0 = 100000.;
+
+}
